requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
pydantic-settings
apscheduler